Fix ICMP type in is_networkacls
In the is_networkacls resource, ICMP type was incorrectly being set
to the value of the ICMP code field. This results in the state being
continually out of sync with resources when ICMP type is set to
a different value to code.

This commit sets the type correctly and modifies the networkacl tests
to validate this. Without the fix, the tests fail because Terraform
creates a non-empty plan after the initial test step that
creates the `is_networkacls` resource.

Fixes #2559