Relocating all power constants to one file
Most of the argument/attribute names are defined in an external helpers package.
Some of these constants are currently being defined in specific data_source files.

This PR moves the argument/attirbute constants defined in data_source files to a singular constant file.
In the future, more PR's will move the constants in the external package into this new constant file.

Files affected
 - data_source_ibm_pi_keys.go
 - data_source_ibm_pi_sap_profiles.go
 - resource_ibm_pi_instance.go
 - resource_ibm_pi_placement_group.go
 - resource_ibm_pi_volume.go
 - resource_ibm_pi_vpn_connection.go