fix: pass ibm cloud session token along to iaas
Currently, in order to use the IBM provider to manager classic
infrastructure resources you must either configure the provider manually
with `iam_token` and `iam_refresh_token` variables, or you must provide
api-specific credentials via the `iaas_classic_username` and
`iaas_classic_api_key` variables. However, it should be reasonable when
using `ibmcloud_api_key` for the provider to passthrough the access and
refresh tokens received as part of the IBM Cloud session.

It seemed like the original code had been in-place to do that, but it
was incorrectly guarded by a non-nil check of
`sess.SoftLayerSession.IAMToken` rather than a check to confirm that
explicit iaas credentials hadn't been provided already.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>