Fix ibm_pi_instance update in place without reboot
PowerVS VSI update in-place with 'ibm_pi_instance' timed out when VSI
status is 'SHUTDOWN' because it was using the
'isWaitForPIInstanceAvailable' poll (but the VSI will never reach
'ACTIVE' state without some external operation). When an
'ibm_pi_instance' resource can be updated in-place _without_ a reboot
the final "isWaitFor" poll should be targeting both 'ACTIVE' and
'SHUTOFF' states (since the VSI will return to it's starting state
after a successful 'RESIZE').