Fix ibm pi instance update wait for status (#4259)
* Fix target state info in wait for update log

The target state is either 'ACTIVE' or 'SHUTOFF', but only 'SHUTOFF' was
mentioned in the log message. This commit adds 'ACTIVE' to the log
message.

* Fix ibm_pi_instance update in place without reboot

PowerVS VSI update in-place with 'ibm_pi_instance' timed out when VSI
status is 'SHUTDOWN' because it was using the
'isWaitForPIInstanceAvailable' poll (but the VSI will never reach
'ACTIVE' state without some external operation). When an
'ibm_pi_instance' resource can be updated in-place _without_ a reboot
the final "isWaitFor" poll should be targeting both 'ACTIVE' and
'SHUTOFF' states (since the VSI will return to it's starting state
after a successful 'RESIZE').