feat: wrap provider to create new error structures
This commit adds code that wraps the provider in order to catch errors and
re-format them with the new system, to minimize the surface area of the
changes required for non-generated resource/data source code.

The primary change to resource code comes in the from of updating calls
to `fmt.Errorf` to be calls to a new wrapper function, `flex.FmtErrorf`.
It accepts the same arguments but instead of wrapping errors coming from
the underlying Go SDK so that we retain only the message, it creates
instances of the new TerraformProblem type to be propagated to the wrapped
provider code. This way, we preserve all of the information needed for the
new error formats, including the information from previous errors.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>