Replace forcenew with diffsuppressfunc
The previous solution replaced the list with a set to allow
terraform to ignore changes in the order of networks during a read,
but if networks were to get added via something like
ibmcloud_pi_network_port_attach then this currently triggers the
forcenew. This doesn't make sense because a user is allowed to
attach networks or volumes without having to create a new instance.