fix: add helper for discriminating terraform problems
If multiple errors may occur for the same operation in the same resource,
and they are not caused by other tools like the SDK, we need a way to
make the IDs for each error scenario unique. The way we support to do this
is with a "discriminator" value but the `TerraformErrorf` function doesn't
expose a way to set the discriminator. This commit adds an additional helper
for use in scenarios where we need to create a discriminated error.

Signed-off-by: Dustin Popp <dustinpopp@ibm.com>