Standardize Arg_CloudInstanceID argument across all data source and resource (#6249)
There were a couple of data sources that did not have the
validation.NoZeroValues check that is required, and force new
was added to the resources. The resources need force new on
this argument because a user could accidentally update that
field and think it does something when in reality it just
messes up the state. Data sources don't need the value
because data sources don't do updates.